android notif [nfc]: Pull channel-handling into its own file
